Develop (#145)
* minor: add NumberOfConsecutiveAppointmentsRequired to bookings availability requests

* chore: version bump

* minor: add NumberOfConsecutiveAppointmentsRequired to bookings availability requests

* chore: version bump

* minor: adds and refactors booking requests to enable Number of consecutive appointments being booked

* patch: some minor refactor and breaking booking request to it's own file

* feat(Stockport Can Fund): Add mail entity

* minor: add shared models for formbuilder post payment
