+++ Models/Booking/Request/AvailabilityRequest.cs
@@ -1,29 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace StockportGovUK.NetStandard.Models.Booking.Request
 {
-    public class AvailabilityRequest
+    public class AvailabilityRequest : BaseRequest
     {
-        public AvailabilityRequest()
+        public AvailabilityRequest() : base()
         {
-            NumberOfConsecutiveAppointmentsRequired = 1;
         }
-
-        [Required]
-        public Guid AppointmentId { get; set; }
 
         [Required]
         public DateTime StartDate { get; set; }
 
         [Required]
         public DateTime EndDate { get; set; }
-
-        public int NumberOfConsecutiveAppointmentsRequired {get; set; }
-
-        public List<BookingResource> OptionalResources { get; set; }
-
-
     }
 }
