+++ Models/Booking/Request/BookingResource.cs
@@ -1,0 +1,14 @@
+using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace StockportGovUK.NetStandard.Models.Booking.Request
+{
+    public class BookingResource
+    {
+        [Required]
+        public Guid ResourceId { get; set; }
+
+        [Required]
+        public int Quantity { get; set; }
+    }
+}
