+++ Models/FormBuilder/EPaymentStatusEnum.cs
@@ -1,0 +1,10 @@
+namespace StockportGovUK.NetStandard.Models.FormBuilder
+{
+    public enum EPaymentStatus
+    {
+        Success = 0,
+        Failure = 1,
+        Cancelled = 2,
+        Declined = 3
+    }    
+}
