+++ Models/StockportCanFund/StockportLocalFundMailModel.cs
@@ -1,0 +1,9 @@
+﻿using StockportGovUK.NetStandard.Models.Mail;
+
+namespace StockportGovUK.NetStandard.Models.StockportCanFund
+{
+    public class StockportCanFundMailModel : BaseMailModel
+    {
+        public string Reference { get; set; }
+    }
+}
